+++ src/test/kotlin/com/github/themartdev/intellijgleam/annotating/GleamHighlightingTest.kt
@@ -1,0 +1,10 @@
+package com.github.themartdev.intellijgleam.annotating
+
+import org.junit.Test
+
+class GleamHighlightingTest : GleamHighlightingTestBase() {
+    override fun getTestDataPath() = "src/test/testData/highlighting"
+
+    @Test
+    fun testFunctionCall() = testHighlighting("function_call.gleam")
+}